Add Node Resolver
